Favorites: DataProvider added - StSDK calls it's methods to work with data
